feat: :clown_face: Add data.js with a static array
The data.js file was added to the project which contains a static array of integers.
This data is intended to be used for testing the functionalities of different algorithms in the application.
The variety of numbers allows us to create different scenarios and validate our algorithms under different conditions.