feat: Add return statement to dequeue method and expose getAll function
Updated the 'dequeue' method to return the removed item and provided a new 'getAll' method for accessing entire queue in 'queue.js'. The queue object is also exposed now to be used in other modules with 'module.exports'. This was done to improve the flexibility and usability of the queue.