feat: Add Integer to English Words converter
This commit introduces a new functionality to convert a non-negative integer to its English words representation. The new converter is implemented in the `integer-to-english-words.js` file. The converter utilizes a Mapped data structure to store a range of English words for numeric value pairs for optimal lookup. Test samples are included in the new script file to validate the functionality of the converter.

In addition to this, minor changes to `integer-to-roman.js` are included to update the way numbers are outputted for easier readability.