feat: Add function to find the median of two sorted arrays
Implemented a function `findMedianSortedArrays` to calculate the median of two sorted lists following the constraints. Note that some portions of the code are yet to be optimized, such as limiting the total length to under 2000 and constraining the number range to [-10^6, 10^6]. This initial implementation uses concatenation of two arrays and sorts them. However, it does not make use of the fact that the given inputs are already sorted. For any improvements, consider implementing optimization methods that utilize the pre-sorted state of input arrays.